move check reg to utils, add tests
